refactor(utils/components): do not use imports from '/pages/* [YTFRONT-3814]'
