+++ packages/ui/src/ui/pages/components/tabs/nodes/tables.scss
