+++ packages/ui/src/ui/store/selectors/components/nodes/nodes/data.ts
@@ -18,7 +18,7 @@ import {
     PropertiesByColumn,
     defaultColumns,
     getNodeTablesProps,
-} from '../../../../../utils/components/nodes/tables';
+} from '../../../../../pages/components/tabs/nodes/tables';
 import {
     getComponentNodesFilterPredicates,
     getComponentNodesFilterStatePredicate,
